+++ R/GLupdater.R
@@ -1,332 +1,317 @@
-#GLupdater v1.0.0 17AUG24 R. Nickens & S.J. Mack
+#GLupdater v2.0.0 27FEB2025 R. Nickens & S.J. Mack
 
-################
-##updateGL
-#'Update a GL String Code to a Specified IPD-IMGT/HLA Database Version.
-#'
-#'A quality control wrapper for GLupdate, which updates a GL String Code to a desired reference database version.
-#'
-#'@param GLStringCode A character string of HLA allele names and operators in GL String Code format, signifying their relation with one another and the associated IMGT/HLA Database release version.
-#'@param Version A character string identifying of the desired IPD-IMGT/HLA Database release version to which the alleles should be updated, going back to version 1.05.0
-#'@param expand A logical that indicates whether user would like to return all allele names that contain the input allele name (TRUE), or if only the direct HLA ID match should be returned (FALSE).
-#'
-#'@return A version of the input GL String code (in the form of a character string) updated to the desired version.
-#'
-#'@export
-#'
-#'@examples
-#'updateGL("hla#1.05.0#HLA-DPA1*0106", "3.52.0")
-#'updateGL("hla#3.36.0#HLA-B*15:35", "3.52.0")
-#'updateGL("hla#3.45.0#HLA-DPA1*02:01:01:04", "3.52.0")
-#'\donttest{
-#'updateGL("hla#3.45.0#HLA-A*02:08", "3.52.0", expand = TRUE)
-#'updateGL("hla#1.05.0#HLA-DPA1*0106", "3.52.0", expand = TRUE)
-#'updateGL("hla#1.05.0#HLA-DPA1*0106", "2.27.0", expand = TRUE)
-#'}
-#'@references Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
-#'@references Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
-updateGL <- function(GLStringCode, Version, expand = FALSE) {
-  #making sure desired version is possible
-  outpos <- GLV2(Version)
-  if(substr(outpos[1],1,1) == "X") {
-  #checking GL String code input, making changes if necessary
-  GLString <- GLvalidate(GLStringCode)
-  #if code is not stopped by previous two lines of code, call GLupdate
-  if(!GLString == FALSE){
-    return(GLupdate(GLString = GLString, Version = Version, expand = expand))
-            } 
-  } else {
-    return(outpos) }
-}
-
-################
+###############
 ##multiUpdateGL
-#'Update a column of GL String Code data to a desired IPD-IMGT/HLA Database version.
+#'Update a Vector of GL String Code Data to a Desired IPD-IMGT/HLA Database Version
 #'
 #'Updates a column from a data frame in GL String Code format to a desired reference database version.
 #'
-#'@param GLstringArray An array of HLA allele names and operators in GL String code format identifying their relation with one another and the pertinent IPD-IMGT/HLA Database release version.
-#'@param Version A character string identifying the desired version to which the alleles should be updated, going back to version 1.05.0.
-#'@param expand A logical to determine whether to include only the direct HLA ID match, or all possible allele matches.
+#'@param GLSCs A vector of GL String codes. 
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'GLSCs' to. Values can range from version 1.05.0 to the loaded version of the alleleListHistory object.
+#'@param expand A logical that indicates if the lowest-numbered truncated allele names that match truncated allele names in 'GLSCs' should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding the update process should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
 #'
-#'@return A version of the input array of GL String Codes (in a data frame) updated to the desired version. NA values are returned in place of alleles that are not present in Version.
+#'@return A version GLSCs (in a data frame) updated to the desired version. NA values are returned in place of alleles that are not present in 'to'.
 #'
 #'@export
 #'
 #'@examples
-#'\donttest{
-#' # Example update of two GL Strings Codes containing truncated alleles from version 3.1.0 to 3.53.0.
-#'GLSC.ex[[2]][1:2] 
-#'multiUpdateGL(GLSC.ex[[2]][1:2], Version = "3.53.0")
-#'}
-#'
-#'@references Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
-#'@references Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
-multiUpdateGL <- function(GLstringArray, Version, expand = FALSE) {
-  #making sure desired version is possible
-      outpos <- GLV2(Version)
-      names <- colnames(GLstringArray)
-      retArray <- data.frame(y = 1:nrow(as.data.frame(GLstringArray)))
-      colnames(retArray) <- colnames(as.data.frame(GLstringArray))
-          
-      if (is.null(names)) {
-            GLstringArray <-as.data.frame(GLstringArray)
-        }
-          names <- colnames(GLstringArray)
+#'multiUpdateGLs <- multiUpdateGL(GLSC.ex$GL.String.Code[1:50],"2.25.0")
+#'multiUpdateGLs <- multiUpdateGL(GLSC.ex$GL.String.Code[1:50],"3.58.0")
+multiUpdateGL <- function(GLSCs,to,expand=FALSE,verbose=FALSE){
   
-          codeSplit <- as.data.frame(do.call(rbind,strsplit(GLstringArray[,1],"#",fixed = TRUE)))
-          allVersions <- unique(codeSplit$V2)
-          
-          stringBlocks <- as.list(rep("",length(allVersions))) ## 1 list element for each release version 1-N
-          names(stringBlocks) <- allVersions ## new; each list element is named for the version 
-          uniqueVec <- vector()
+  if(!checkVersion(to)) {stop(paste("Release version",to,"is not loaded in the HLAtools package. Please provide a loaded release version for 'to', or update the alleleListHistory object."))}
+  if(!is.logical(expand)) {stop(paste(expand,"is not a valid value for 'expand'. Please provide a logical value."))}
+  if(!is.logical(verbose)) {stop(paste(verbose,"is not a valid value for 'verbose'. Please provide a logical value."))}
   
-          for(i in 1:length(allVersions)){ ## a vector of the release version names; each i is a relesae vesion
-            
-              stringBlocks[[i]] <- as.numeric(rownames(codeSplit[codeSplit$V2 == allVersions[i],]))
+  outVec <-rep(NA,length(GLSCs))
+  
+  for(i in 1:length(GLSCs)){
     
-                for(j in 1:length(stringBlocks[[i]])) {
-                                                                                                               ## vvvv 1 instead of j    
-                  uniqueVec <- unique(c(uniqueVec,strsplit(codeSplit$V3[stringBlocks[[i]][j]],split="[+?~+/^\\|]")[[1]])) 
-                }
-
-              uniqueVec <- sort(uniqueVec)
-              
-              vers <- allVersions[i] ## fixing improperly structured vers values e.g., 3.1.0 instead of 3.01.0
-              if(!validateVersion(vers)) {
-                vers <- expandVersion(substr(GLV2(vers),2,stop = nchar(GLV2(vers))))
-              }
-              
-              if(expand){ ## creating the lookup table for returned strings
-                    lookUp <- as.data.frame(cbind(uniqueVec,paste("HLA-",suppressMessages(GIANT(uniqueVec,vers,Version)),sep=""))) ## silently call GIANT when expand = TRUE
-                    } else {
-                    lookUp <- as.data.frame(cbind(uniqueVec,paste("HLA-",GIANT(uniqueVec,vers,Version),sep=""))) 
-                }
-              colnames(lookUp) <- c("from","to")
-              
-              ### When expand = TRUE; apply queryRelease() to find all of the names that match in release X
-              if(expand){
-                    lookUp <- cbind(lookUp,as.data.frame(rep(NA,nrow(lookUp))))
-                    colnames(lookUp)[3] <- "multi"
-                    
-                    for(j in 1:nrow(lookUp)) {
-                      
-                            only <- queryRelease(Version,gsub("HLA-","",lookUp$from[j],fixed = TRUE),TRUE) ## all of the string-matches in that release
-                            locPre <- strsplit(gsub("HLA-","",lookUp$from[j],fixed = TRUE),split = "*",fixed = TRUE)[[1]][1] # target locus, for length and match
-                            lookUp$multi[j] <- paste(only[substr(only,1,nchar(locPre)+1) == paste(locPre,"*",sep="")],collapse="/") # exclude non-locus matches
-                            message("Returning all matches for ", strsplit(lookUp$from[j],"-")[[1]][2],".",sep="")
-                            lookUp$multi[j] <- str_replace_all(lookUp$multi[j],"/",paste("/HLA-",sep=""))
-                            lookUp$multi[j] <- paste("HLA-",lookUp$multi[j],sep="")
-                    }
-                }
-              
-             for(j in 1:length(stringBlocks[[i]])) { ## translating for each set of stringBlocks
-                 for(k in 1:nrow(lookUp)) {
-                                                                                                          # switch between multi or single lookup 
-                        codeSplit$V3[stringBlocks[[i]][j]] <- gsub(pattern = lookUp$from[k],replacement = ifelse(expand, lookUp$multi[k],lookUp$to[k]), x = codeSplit$V3[stringBlocks[[i]][j]],fixed = TRUE)
-                                }
-                        codeSplit$V2[stringBlocks[[i]][j]] <- Version
+    outVec[i] <- updateGL(GLSCs[i],to,expand,verbose)
+    
+  }
   
-                        retArray$GLstringArray[stringBlocks[[i]][j]] <- paste(unlist(codeSplit[stringBlocks[[i]][j],]),collapse="#") 
-                      }
-                 }
-          
- return(retArray$GLstringArray)
+  outVec
+}        
+
+###############
+##multiTranslateGLstring
+#'Translate a Vector of GL Strings to a Desired IPD-IMGT/HLA Database Version
+#'
+#'Translates a column from a data frame in GL String Code format to a desired reference database version.
+#'
+#'@param GLstrings A vector of GL Strings encoded in the same IPD-IMGT/HLA Database release version. 
+#'@param from A character string identifying the IPD-IMGT/HLA Database release version of 'GLstrings'.
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'GLstrings' to. Values can range from version 1.05.0 to the loaded version of the alleleListHistory object.
+#'@param expand A logical that indicates if the lowest-numbered truncated allele names that match truncated allele names in 'GLstrings' should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding the trnslation process should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
+#'
+#'@return A version GLSCs (in a data frame) updated to the desired version. NA values are returned in place of alleles that are not present in 'to'.
+#'
+#'@export
+#'
+#'@examples
+#'multiGLs <- multiTranslateGLstring(GLstring.ex$Gl.String[1:50],"3.01.0","2.15.0")
+#'multiGLs <- multiTranslateGLstring(GLstring.ex$Gl.String[1:50],"3.30.0","3.58.0")
+multiTranslateGLstring <- function(GLstrings,from,to,expand=FALSE,verbose=FALSE){
+  
+  if(!checkVersion(to)) {stop(paste("Release version",to,"is not loaded in the HLAtools package. Please provide a loaded release version for 'to', or update the alleleListHistory object."))}
+  if(!checkVersion(from)) {stop(paste("Release version",from,"is not loaded in the HLAtools package. Please provide a loaded release version for 'from', or update the alleleListHistory object."))}
+  if(!is.logical(expand)) {stop(paste(expand,"is not a valid value for 'expand'. Please provide a logical value.\n"))}
+  if(!is.logical(verbose)) {stop(paste(verbose,"is not a valid value for 'verbose'. Please provide a logical value.\n"))}
+  
+  outVec <-rep(NA,length(GLstrings))
+  
+  for(i in 1:length(GLstrings)){
+    
+    outVec[i] <- translateGLstring(GLstrings[i],from,to,expand,verbose)
+    
+  }
+  
+  outVec
+}        
+
+###############
+##updateGL
+#'Update a GL String Code to a Specified IPD-IMGT/HLA Database Version
+#'
+#'Updates the elements of a GLString Code across IPD-IMGT/HLA Database release versions. Truncated allele names can be expanded to the list of all allele names that contain the truncated name.
+#'
+#'@param GLSC A GL String Code containing HLA allele names. 
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'GLSC' to. Values can range from version 1.05.0 to the loaded version of the alleleListHistory object.
+#'@param expand A logical that indicates if the lowest-numbered truncated allele names that match truncated allele names in 'GLSC' should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding the update process should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
+#'
+#'@return A GL String Code updated the 'to' release version.
+#'
+#'@export
+#'
+#'@examples
+#'updateGL(GLSC.ex$GL.String.Code[1],"2.15.0",FALSE,FALSE)
+#'updateGL(GLSC.ex$GL.String.Code[1],"2.15.0",TRUE,FALSE)
+#'
+updateGL <- function(GLSC,to,expand = FALSE,verbose = FALSE) {
+  
+  if(!checkVersion(to)) {stop(paste("Release version",to,"is not loaded in the HLAtools package. Please provide a loaded release version for 'to', or update the alleleListHistory object."))}
+  if(!is.logical(expand)) {stop(paste(expand,"is not a valid value for 'expand'. Please provide a logical value."))}
+  if(!is.logical(verbose)) {stop(paste(verbose,"is not a valid value for 'verbose'. Please provide a logical value."))}
+  
+  parts <- strsplit(GLSC,"#",TRUE)[[1]]
+  fromVers <- parts[2] # -- version to translate from
+  if(!checkVersion(fromVers)) {stop(paste("Release version",fromVers,"is not loaded in the HLAtools package. Please update the alleleListHistory object."))}
+  
+  inString <- parts[3] # -- GL string to translate to 
+  outString <- rep(NA,length(inString)) ## -- the translated string 
+  
+  outString <- translateGLstring(inString,fromVers,to,expand,verbose)
+  
+  paste("hla",to,outString,sep="#")
   
 }
 
+###############
+##GLupdate
+#'Update a GL String Code to a Specified IPD-IMGT/HLA Database Version
+#'
+#'Updates the elements of a GLString Code across IPD-IMGT/HLA Database release versions. Truncated allele names can be expanded to the list of all allele names that contain the truncated name.
+#'
+#'@param GLSC A GL String Code containing HLA allele names. 
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'GLSC' to. Values can range from version 1.05.0 to the loaded version of the alleleListHistory object.
+#'@param expand A logical that indicates if the lowest-numbered truncated allele names that match truncated allele names in 'GLSC' should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding the update process should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
+#'
+#'@export
+#'
+#'@return A GL String Code updated the 'to' release version.
+#'
+#'@note This function is maintained for compatibility with previous releases, and will be removed in a future release. HLAtools users are advised to use updateGL().
+#'
+#'@examples
+#'GLupdate(GLSC.ex$GL.String.Code[1],"2.15.0",FALSE,FALSE)
+#'GLupdate(GLSC.ex$GL.String.Code[1],"2.15.0",TRUE,FALSE)
+#'
+GLupdate <- function(GLSC,to,expand = FALSE,verbose = FALSE) {
+  
+  updateGL(GLSC,to,expand,verbose)
+  
+}
+###############
+##translateGLstring
+#'Translate a GL String across IPD-IMGT/HLA Database Release Versions
+#'
+#'Translates a single Genotype List (GL) string across IPD-IMGT/HLA Database release versions. Truncated allele names can be expanded to the list of all allele names that contain the truncated name.  
+#'
+#'@param GLstring A GL String of HLA allele names. 
+#'@param from A character string identifying the IPD-IMGT/HLA Database release version of 'GLstring'.
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'GLstring' to. Values can range from version 1.05.0 to the loaded version of the alleleListHistory object.
+#'@param expand A logical that indicates if the lowest-numbered truncated allele names that match truncated allele names in the GL String should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding allele translations should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
+#'
+#'@export
+#'
+#'@return A GL String of allele names in the 'to' release version. If the GL String is determined to be invalid, FALSE is returned. 
+#'
+#'@examples
+#'translateGLstring(GLstring.ex$Gl.String[1],"3.01.0","2.15.0")
+#'translateGLstring(GLstring.ex$Gl.String[1],"3.01.0","2.15.0",TRUE)
+#'translateGLstring(GLstring.ex$Gl.String[1],"3.01.0","2.15.0",FALSE,TRUE)
+#'translateGLstring(GLstring.ex$Gl.String[1],"3.01.0","2.15.0",TRUE,TRUE)
+#'
+translateGLstring <- function(GLstring,from,to,expand=FALSE,verbose=FALSE) {      
+  
+  if(!checkVersion(to)) {stop(paste("Release version",to,"is not loaded in the HLAtools package. Please provide a loaded release version for 'to', or update the alleleListHistory object."))}
+  if(!checkVersion(from)) {stop(paste("Release version",from,"is not loaded in the HLAtools package. Please provide a loaded release version for 'from', or update the alleleListHistory object."))}
+  if(!is.logical(expand)) {stop(paste(expand,"is not a valid value for 'expand'. Please provide a logical value.\n"))}
+  if(!is.logical(verbose)) {stop(paste(verbose,"is not a valid value for 'verbose'. Please provide a logical value.\n"))}
+  
+      if(verbose) {
+        if(!validateGLstring(GLstring,"1.1")) {return(FALSE)}
+        } else {
+          if(suppressMessages(validateGLstring(GLstring,"1.1")) == FALSE) {return(FALSE)}
+        }
+  
+        #set up all of the parts
+          bits <- strsplit(GLstring,"[+,/,|,?,^,~]")[[1]] ## the alleles to be translated, in order of appearance 
+          
+            if(length(bits) == 1) {
+              return(translateAllele(bits,from,to,expand,verbose))
+                  }
+          
+          pieces <- rep(NA,length(bits)-1) ## the operators that delimit the alleles, in order.
+          nuts <- ifelse(grep("HLA-",GLstring),TRUE,FALSE) ## determine if the returned vector elements should be prefixed with "HLA-".
+          bolts <- 0 ## spacers for determining the nature and order of delimiters
+          screws <- rep(NA,length(bits)) ## the new, translated allele/strings, corresponding to each bit
+  
+          for(i in 1:length(pieces)) { # determine the nature and order of delimiters
+              pieces[i] <- substr(GLstring,bolts+nchar(bits[i])+1,bolts+nchar(bits[i])+1)
+              bolts <- bolts+nchar(bits[i])+1
+            }
+           
+            for(i in 1:length(bits)) { ## looping through each allele in the string
+              screws[i] <- translateAllele(bits[i],from,to,expand,verbose)
+                } 
+          
+            paste(paste(c(rbind(screws[1:length(screws)-1],pieces)),collapse=""),screws[length(screws)],sep="")
+                
+}
 
 ################
-##GLupdate
-#'Update a GL String Code.
+##translateAllele
+#'Translate HLA Allele Names Across IPD-IMGT/HLA Database Release Versions
 #'
-#'Updates the allele names in a Genotype List (GL) String Code to the desired reference database version using the IPD-IMGT/HLA Database's Allele List History table.
+#'Translates a single HLA allele name across IPD-IMGT/HLA Database release versions.  Truncated allele names in one version can be expanded to a list of all allele names that match the provided allele name in a chosen version.
 #'
-#'@param GLString A character string of HLA allele names and operators in GL String Code format signifying their relation with one another and the pertinent HLA Allele List version.
-#'@param Version A character string of the desired version to which the alleles to be updated, going back to version 1.05.0.
-#'@param expand A logical that determines whether to return only the direct HLA ID allele match or all possible HLA allele matches.
+#'@param allele A full or truncated HLA allele name. 
+#'@param from A character string identifying the IPD-IMGT/HLA Database release version of 'allele'.
+#'@param to A character string identifying the IPD-IMGT/HLA Database release version to translate 'allele' to. Values can range from version 1.05.0 to the loaded version of the the alleleListHistory object. 
+#'@param expand A logical that indicates if the lowest-numbered truncated allele name that matches a truncated allele name should be returned (expand = FALSE), or if a slash-delimited string of all matching full-length allele names should be returned (expand = TRUE). The default value is FALSE.
+#'@param verbose A logical that indicates if messages regarding allele translations should be sent to the console (TRUE) or not (FALSE). The default value is FALSE.
 #'
-#'@return An updated version the GL String Code input (in the form of a character string) updated to the input desired version.
-#'
-#'@note For internal use only.
-#'
-#'@importFrom stringr fixed str_replace_all
+#'@return A character string of either a single matching allele or all matching alleles in the 'to' release version. If an allele does not have a cognate match in 'to', NA is returned for that allele. 
 #'
 #'@export
-#'
+#'         
 #'@examples
-#'GLupdate("hla#3.36.0#HLA-B*15:35", "3.52.0")
-#'GLupdate("hla#3.45.0#HLA-A*02:08", "3.52.0")
-#'GLupdate("hla#3.45.0#HLA-A*02:08", "3.52.0", expand = TRUE)
-#'\donttest{
-#'GLupdate("hla#3.1.0#HLA-A*02:01+HLA-A*01:01:01:01", Version = "3.53.0")
-#'GLupdate("hla#3.44.0#HLA-DPA1*02:01:01:05", "3.45.0")
-#'GLupdate("hla#3.45.0#HLA-DPA1*02:01:01:05", "3.46.0")
-#'GLupdate("hla#3.37.0#HLA-A*01:02", "3.52.0", expand = TRUE)
-#'}
+#'translateAllele("A*01:01","3.01.0","2.20.0",FALSE)
+#'translateAllele("A*0101","2.20.0","3.01.0",TRUE)
+#'translateAllele("B*57:01","3.12.0","3.12.0",TRUE)
 #'
-#'@references Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
-#'@references Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
-#'@source https://github.com/ANHIG/IMGTHLA/blob/Latest/Allelelist_history.txt
-#'
-GLupdate <- function(GLString, Version, expand = FALSE) {
-  #getting input and output versions formatted correctly
-  inpos <- GLV(GLString)
-  outpos <- GLV2(Version)
-  #this is to determine if a select group of genes need to be treated differently when updating from very old versions to recent versions
-  outnum <- gsub("X", "",outpos)
-  #creating string to return back
-  gBack <- GLString
-
-  #creating list of all alleles
-  gBack <- gsub("HLA-","",gBack)
-  origV <- strsplit(GLString, "#")[[1]][2]
-  nameList <- strsplit(gBack, "#")[[1]][3]
-  nameList <- strsplit(nameList, "[.]|[/]|[?]|[+]|[|]|[~]|\\^")[[1]]
-  for (i in 1:length(nameList[[1]])) {
-    nameList[[1]][i] <- gsub("HLA-","",nameList[[1]][i])
+translateAllele <- function(allele,from,to,expand=FALSE,verbose = FALSE) {
+  
+  if(!checkVersion(to)) {stop(paste("Release version",to,"is not loaded in the HLAtools package. Please provide a loaded release version for 'to', or update the alleleListHistory object."))}
+  if(!checkVersion(from)) {stop(paste("Release version",from,"is not loaded in the HLAtools package. Please provide a loaded release version for 'from', or update the alleleListHistory object."))}
+  if(!is.logical(expand)) {stop(paste(expand,"is not a valid value for 'expand'. Please provide a logical value.\n"))}
+  if(!is.logical(verbose)) {stop(paste(verbose,"is not a valid value for 'verbose'. Please provide a logical value.\n"))}
+  
+  nuts <- ifelse(substr(allele,1,4) == "HLA-", TRUE,FALSE) ## should the returned vector elements be prefixed with "HLA-".
+  
+  origField <- origDigit <- NA ## determine if 'allele' is field or digit delimited, and how many fields
+  if(length(grep(":",allele)) == 0) {
+    origDigit <- nchar(strsplit(allele,"*",fixed = TRUE)[[1]][2])  #vs 1 & 2 allele names
+  } else { origField <- length(strsplit(allele,":",fixed=TRUE)[[1]]) # v3 allele names
   }
-  #ensuring no replicates
-  nameList <- unique(nameList)
-
-  if(expand) {
-    for (i in 1:length(nameList)) {
-      #list to put all name matches in for an allele
-      nameOp <- vector("list", 1)
-      rTU <- ""
-      backBack <- nameList[i]
-      #length of namelist value, will be used to shorten items in AlleleListHistory and to decide how to compare value
-      LX <- length(strsplit(as.character(nameList[i]), "[:]|[*]")[[1]])
-      #specific case when updating very old versions to recent version, must manually add ":" to get matches
-      hope <- strsplit(strsplit(nameList[i], "[*]")[[1]][2], "")[[1]]
-      #version 2.28.0 was last version not to use ":"
-      if (outnum > 2280) {
-        if (LX ==2) {
-          #formatting input correctly so it may accurately be compared to the new version
-          if (length(hope)==4) {
-            xxx <- paste0(paste0(paste0(hope[1], hope[2]), ":"), paste0(hope[3]), hope[4])
-            yyy <- paste0(paste0(strsplit(nameList[i], "[*]")[[1]][1], "*"), xxx)
-            nameList[i] <- yyy
-          }
-          if (length(hope)==5) {
-            xxx <- paste0(paste0(paste0(hope[1], hope[2]), ":"), paste0(hope[3]), hope[4])
-            yyy <- paste0(paste0(strsplit(nameList[i], "[*]")[[1]][1], "*"), xxx)
-            yyy <- paste0(paste0(yyy, ":0"), hope[5])
-            nameList[i] <- yyy
-          }
-          if (length(hope)==6) {
-            xxx <- paste0(paste0(paste0(paste0(hope[1], hope[2]), ":"), paste0(paste0(hope[3], hope[4]), ":")), paste0(hope[5], hope[6]))
-            yyy <- paste0(paste0(strsplit(nameList[i], "[*]")[[1]][1], "*"), xxx)
-            nameList[i] <- yyy
-          }
-        }
-      }
-      LX <- length(strsplit(as.character(nameList[i]), "[:]|[*]")[[1]])
-      #search everything in column for match
-      for (x in 1:length(alleleListHistory$AlleleListHistory[[outpos]])) {
-        #original value in box being searched, will be added to list if altered allelename matches with input nameslist
-        YORG <- alleleListHistory$AlleleListHistory[[outpos]][x]
-        #creating altered allelename for specific item in AlleleListHistory
-        YWAY <- ""
-        YWA <- strsplit(as.character(alleleListHistory$AlleleListHistory[[outpos]][x]), "[:]|[*]")[[1]]
-        try2 <- ""
-        #this will only be used in older versions without decimals
-        if (LX == 2) {
-          hope2 <- strsplit(YWA[2], "")[[1]]
-          if (length(hope2) > length(hope)) {
-            te <- ""
-            for (w in 1:length(hope)) {
-              te <- paste0(te, hope2[w])
-            }
-            YWA[2] <- te
-          }
-          YWAY <- paste0(paste0(YWA[1], "*"), YWA[2])
-        }
-        #most inputs will go through this route
-        if (LX > 2) {
-          for (n in 2:(LX-1)) {
-            YWAY <- paste0(paste0(YWAY, YWA[n]), ":")
-          }
-          YWAY <- paste0(paste0(YWA[1], "*"), YWAY)
-          YWAY <- paste0(YWAY, YWA[LX])
-        }
-        #if input allelename and altered allelename from AllelelistHistory match, add original name from AlleleListHistory to the list of matches
-        if(!is.na(YWAY)&!is.na(nameList[i])) {
-          if(YWAY == nameList[i]){
-            nameOp[[1]] <- append(nameOp[[1]], paste0("HLA-", YORG))
-          }
-        }
-
-      }
-      nameOp <- unique(nameOp)
-
-      if (length(nameOp[[1]]) > 1) {
-        for (w in 1:(length(nameOp[[1]])-1)) {
-          rTU <- paste0(rTU, paste0(nameOp[[1]][w], "/"))
-        }
-        rTU <- paste0(rTU, nameOp[[1]][length(nameOp[[1]])])
-      }
-      if (length(nameOp[[1]]) ==1) {
-        rTU <- nameOp[[1]][1]
-      }
-
-      if(is.null(nameOp[[1]])) {
-        message(nameList[i], " has not been located in desired GL String Version.")
-        gBack <- str_replace_all(gBack, fixed(backBack), "NA")
-      }
-      #editing string to return
-      if (!is.null(nameOp[[1]])) {
-        message("Returning all matches for ", nameList[i],".",sep="")
-        gBack <- str_replace_all(gBack, fixed(backBack), rTU)
-      }
-
-    }
-    #changing version to reflect the update
-    gBack <- gsub(origV,Version,gBack, fixed = TRUE) ### v1.1.4 adding fixed = TRUE fixes the issue
-    gBack #### this should be a return 1.1.4
+  
+  #---- single allele translation
+  loc <- strsplit(stringr::str_sub(allele,-(nchar(allele)-ifelse(nuts,4,0))),"*",fixed = TRUE)[[1]][1]
+  check <- queryRelease(from,stringr::str_sub(allele,-(nchar(allele)-ifelse(nuts,4,0))),TRUE) # the set of all allele names matching the current allele in the current release
+  check <- sort(check[substr(check,1,nchar(loc)) == loc]) ## exclude alleles from other loci, sort in numerical order
+  ## exclude partial-match-loci. e.g, "DMA*01:01" when the allele is "A*01:01"
+  
+  if(length(check) == 0) {
+    if(verbose) {message(paste("Allele",stringr::str_sub(allele,-(nchar(allele)-ifelse(nuts,4,0))),"is not found in release",paste(from,".",sep=""),sep=" "))}
+    currFrom <- NA
   } else {
-    for (i in 1:length(nameList)) {
-      if(nameList[i] %in% alleleListHistory$AlleleListHistory[[inpos]]){
-        #finding which row the allele is in
-        ytemp <- which(alleleListHistory$AlleleListHistory[[inpos]] == nameList[i])
-        #getting value of allele from desired version column
-        valtemp <- alleleListHistory$AlleleListHistory[ytemp,outpos]
-        #avoiding issues with NA
-        if(is.na(valtemp)|is.null(valtemp)) {
-          message(nameList[i], " has been removed from the Allele List or has had its name changed.")
-          gBack <- str_replace_all(gBack, fixed(nameList[i]), "NA")
-        }
-        #if valtemp is not NA, replace
-        else {
-          #replacing words on gBack string to return
-          gBack <- str_replace_all(gBack, fixed(nameList[i]), paste0("HLA-", valtemp))
-        }
-      }  else {
-        message(nameList[i], " is not present in IPD-IMGT/HLA Database Version ", strsplit(GLString, "#")[[1]][[2]], ".")
-        tempFind <- paste0(c("hla", origV, nameList[i]), collapse = "#")
-        suppressMessages(lastEffort <- GLupdate(tempFind, Version = origV, expand = TRUE))
-        cOne <- strsplit(lastEffort, split = "#",fixed=FALSE)[[1]][3] ### 1.1.4 fix - explicitly set fixed = FALSE
-        cOne <- strsplit(cOne, "[/]",fixed = FALSE)[[1]][1] ### 1.1.4 fix - explicitly set fixed = FALSE
-        if (cOne != "NA") {
-          message("Returning first match for ", nameList[i],".",sep="")
-          Q <- paste0(c("hla", origV, cOne), collapse = "#")
-          temp <- GLupdate(Q, Version = Version)
-          temp <- strsplit(temp, split = "#",fixed=FALSE)[[1]][3] ### 1.1.4 fix - explicitly set fixed = FALSE
-          gBack <- str_replace_all(gBack, fixed(nameList[i]), temp)
-        }
-        else { gBack <- str_replace_all(gBack, fixed(nameList[i]), "NA")
-        }
+    if(length(check) > 1) {
+      if(verbose) {message(paste(length(check), "alleles matching",paste(ifelse(nuts,"HLA-",""),stringr::str_sub(allele,-(nchar(allele)-ifelse(nuts,4,0))),sep=""),"are found in release",paste(from,".",sep=""),"Translation will be performed using the", paste(ifelse(nuts,"HLA-",""),check[1],sep=""),"allele.",sep=" "))}
+      currFrom <- check[1]
+    } else { ## there is only one allele
+      currFrom <- check 
+    }
+  }
+  
+  if(is.na(currFrom)) {  # no currFrom -- return NA
+    return(NA)
+  } else { # there is a currFrom
+    
+    ## determine which row the current allele is in
+    currRow <- match(currFrom,alleleListHistory$AlleleListHistory[,colnames(alleleListHistory$AlleleListHistory) == GLV2(from)])
+    
+    updAllele <- alleleListHistory$AlleleListHistory[currRow,GLV2(to)] ## only need the exact match 
+    
+    if(is.na(updAllele)) {
+      if(verbose){paste("The",ifelse(nuts,paste("HLA",currFrom,sep="-"),currFrom),"allele has no cognate in release",paste(to,".",sep=""),sep=" ")}
+      return(NA)
+    } ## allele is not found in 'to'.
+    
+    updDigit <- updField <- NA ## determine if the updated allele is field or digit delimited
+    
+    if(length(grep(":",updAllele)) == 0) {
+      updDigit <- nchar(strsplit(updAllele,"*",fixed = TRUE)[[1]][2])  #vs 1 & 2 allele names
+    } else { 
+      updField <- length(strsplit(updAllele,":",fixed=TRUE)[[1]]) # v3 allele names
+    }
+    
+    if(!is.na(origField)) { ## original allele was colon-field delimited
+      if(!is.na(updField)) { ## updated allele was colon-field delimited
+        updAllele <- alleleTrim(allele = updAllele,resolution = origField,version = as.numeric(strsplit(to,split = ".",fixed = TRUE)[[1]][1]))
+      } else { ## updated allele was digit-delimited
+        updAllele <- alleleTrim(allele = updAllele,resolution = origField,version = as.numeric(strsplit(to,split = ".",fixed = TRUE)[[1]][1]))
+      }
+    } else { ## original allele was digit delimited
+      if(!is.na(updField)) { ## updated allele was field-delimited
+        updAllele <- alleleTrim(allele = updAllele,resolution = ceiling(origDigit/2),version = as.numeric(strsplit(to,split = ".",fixed = TRUE)[[1]][1]))
+      } else { ## updated field was digit-delimited
+        updAllele <- alleleTrim(allele = updAllele,resolution = ceiling(origDigit/2),version = as.numeric(strsplit(to,split = ".",fixed = TRUE)[[1]][1]))
       }
     }
-    gBack <- gsub(origV,Version,gBack, fixed = TRUE) ### v1.1.4 adding fixed = TRUE fixes the issue
-    gBack
-
-  }
-
+    
+    ## this single allele is what will be returned if expand = FALSE
+    if(!expand) {return(ifelse(nuts,paste("HLA-",updAllele,sep=""),updAllele))}  
+    
+    trans <- updAllele  
+    
+    if(expand) { ## multiple alleles
+      
+      check <- queryRelease(to,trans,TRUE) # the set of all allele names matching the current allele in the current release
+      loc <- strsplit(stringr::str_sub(check[1],-(nchar(check[1])-ifelse(nuts,4,0))),"*",fixed = TRUE)[[1]][1] ## need to do it again for 'C'vs'Cw'
+      updAlleles <- sort(check[substr(check,1,nchar(loc)) == loc]) ## exclude alleles from other loci, sort in numerical order
+      if(verbose) {message(paste(length(updAlleles), "version", to, "alleles match",paste(allele,".",sep=""),sep=" "))}
+      
+      if(nuts){ updAlleles <- paste("HLA-",updAlleles,sep="") }
+      
+      trans <- paste(updAlleles,collapse="/")
+      
+    } ## end of multiple allele expansion
+  } ## end of 'is there a currFrom?'
+  
+  trans # return the translated allele
 }
+######################################################################################
+
 ################
 ##GLV
-#'Retrieve version from input GL String.
+#'Retrieve Version from Input GL String
 #'
 #'Extracts the version data from an input GL String Code, or provides appropriate options if version input is not present in the IPD-IMGT/HLA Database.
 #'
@@ -371,7 +356,7 @@ GLV <- function(GLString) {
 
 ################
 ##GLV2
-#'Format GL String Code version number.
+#'Format GL String Code Version Number
 #'
 #'Returns a compressed IPD-IMGT/HLA Database release version or a vector of dot-delimited release version options if the specified version is not present in the alleleListHistory object.
 #'
@@ -389,7 +374,7 @@ GLV <- function(GLString) {
 #'
 #'@references Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
 #'@references Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
-
+#
 GLV2 <- function(Version) {
   ve1 <- ve2 <- ""
   version <- strsplit(Version, "[.]")[[1]]
@@ -414,9 +399,9 @@ GLV2 <- function(Version) {
 
 ################
 ##GLVhelper
-#'Locate matches for an incomplete IPD-IMGT/HLA Database version.
+#'Locate Matches for an Incomplete IPD-IMGT/HLA Database Version
 #'
-#' Uses the provided version information to locate and return possible matches for an incompletely defined IPD-IMGT/HLA Database release version.
+#'Uses the provided version information to locate and return possible matches for an incompletely defined IPD-IMGT/HLA Database release version.
 #'
 #'@param Version A character string of the desired version to which the alleles should be updated, going back to version 1.05.0.
 #'
@@ -537,7 +522,7 @@ GLVhelper <- function(Version) {
 
 ################
 ##redec
-#'Reintroduce version decimals.
+#'Reintroduce Version Decimals
 #'
 #'Correctly put decimals back into an AlleleListHistory column-formatted version name.
 #'
@@ -573,7 +558,7 @@ redec <- function(Cname) {
 
 ################
 ##GLvalidate
-#'Validates a GL String Code.
+#'Validate a GL String Code
 #'
 #'A lightweight validator that inspects a GL String Code for correct structure. If the namespace field contains a value other than value of 'namespace', the namespace is changed to 'hla'. The version and GL String fields are not evaluated.
 #'